feat: centralize signOut function in AppContext
- Move signOut function from _app.tsx and settings.tsx to AppContext
- Add signOut to AppContext interface and provider
- Update _app.tsx and settings.tsx to use centralized signOut
- Remove duplicate signOut implementations
- Add window.location.reload() for complete page refresh on logout
