+++ frontend/src/pages/settings.tsx
@@ -10,32 +10,11 @@ import Head from 'next/head';
 import { CopyButton } from '@/components/ui/copy-button';
 
 export default function Settings() {
-  const { dispatch, state } = useAppContext();
+  const { state, signOut } = useAppContext();
   const { preference, setThemePreference } = useTheme();
   const [showApiKey, setShowApiKey] = useState(false);
 
   const adminApiKey = state.apiKey || '';
-
-  const signOut = () => {
-    // Clear all localStorage items
-    localStorage.removeItem('payment_api_key');
-    localStorage.removeItem('selectedPaymentSourceId');
-    localStorage.removeItem('userIgnoredSetup');
-    localStorage.removeItem('masumi_last_transactions_visit');
-    localStorage.removeItem('masumi_new_transactions_count');
-    localStorage.removeItem('dialogPosition');
-    localStorage.removeItem('theme');
-
-    // Reset all app state
-    dispatch({ type: 'SET_API_KEY', payload: '' });
-    dispatch({ type: 'SET_PAYMENT_SOURCES', payload: [] });
-    dispatch({ type: 'SET_CONTRACTS', payload: [] });
-    dispatch({ type: 'SET_WALLETS', payload: [] });
-    dispatch({ type: 'SET_RPC_API_KEYS', payload: [] });
-
-    // Force redirect to login page
-    window.location.href = '/';
-  };
 
   return (
     <MainLayout>
